+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @doccov/cli
 
+## 0.25.3
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies
+  - @openpkg-ts/spec@0.12.0
+  - @doccov/sdk@0.25.3
+
 ## 0.25.0
 
 ### Patch Changes