+++ packages/sandbox/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @doccov/sandbox
 
+## 0.1.4
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies
+  - @openpkg-ts/spec@0.12.0
+  - @doccov/sdk@0.25.3
+
 ## 0.1.3
 
 ### Patch Changes